+++ lib/features/kana/presentation/providers/kana_groups_provider.g.dart
@@ -1,0 +1,29 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'kana_groups_provider.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$filteredKanaGroupsHash() =>
+    r'9801048cdfcbd492de92b062cd0bde7a4a402b21';
+
+/// See also [filteredKanaGroups].
+@ProviderFor(filteredKanaGroups)
+final filteredKanaGroupsProvider =
+    AutoDisposeFutureProvider<List<KanaGroup>>.internal(
+      filteredKanaGroups,
+      name: r'filteredKanaGroupsProvider',
+      debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+          ? null
+          : _$filteredKanaGroupsHash,
+      dependencies: null,
+      allTransitiveDependencies: null,
+    );
+
+@Deprecated('Will be removed in 3.0. Use Ref instead')
+// ignore: unused_element
+typedef FilteredKanaGroupsRef = AutoDisposeFutureProviderRef<List<KanaGroup>>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package